Rename Window_internal unit to WindowBase for consistency
