+++ src/openrct2-ui/interface/Window.h
@@ -10,7 +10,7 @@
 #pragma once
 
 #include <openrct2/interface/Window.h>
-#include <openrct2/interface/Window_internal.h>
+#include <openrct2/interface/WindowBase.h>
 
 struct TextInputSession;
