+++ src/openrct2-ui/scripting/ScWindow.hpp
@@ -14,7 +14,7 @@
     #include "ScWidget.hpp"
 
     #include <openrct2/interface/Window.h>
-    #include <openrct2/interface/Window_internal.h>
+    #include <openrct2/interface/WindowBase.h>
     #include <openrct2/scripting/Duktape.hpp>
 
 namespace OpenRCT2::Scripting
