+++ src/openrct2/entity/Peep.cpp
@@ -31,7 +31,7 @@
 #include "../entity/EntityRegistry.h"
 #include "../entity/EntityTweener.h"
 #include "../interface/Viewport.h"
-#include "../interface/Window_internal.h"
+#include "../interface/WindowBase.h"
 #include "../localisation/Formatter.h"
 #include "../localisation/Formatting.h"
 #include "../management/Finance.h"
