+++ src/openrct2/interface/Viewport.cpp
@@ -44,7 +44,7 @@
 #include "../world/tile_element/WallElement.h"
 #include "Colour.h"
 #include "Window.h"
-#include "Window_internal.h"
+#include "WindowBase.h"
 
 #include <cstring>
 #include <list>
