+++ src/openrct2/interface/WindowBase.h
