+++ src/openrct2/scenes/title/Command/SetLocation.cpp
@@ -12,7 +12,7 @@
 #include "../../../OpenRCT2.h"
 #include "../../../interface/Viewport.h"
 #include "../../../interface/Window.h"
-#include "../../../interface/Window_internal.h"
+#include "../../../interface/WindowBase.h"
 #include "../../../world/Map.h"
 
 namespace OpenRCT2::Title
